+++ src/components/molecules/index.ts
@@ -11,6 +11,8 @@ import CalendarModal from "./CalendarModal";
 import Categories from "./Categories";
 import DateInput from "./DateInput";
 import Modal from "./Modal";
+import ImageUpload from "./ImageUpload";
+import ColorPalette from "./ColorPalette";
 
 export {
   SideBar,
@@ -26,4 +28,6 @@ export {
   Categories,
   DateInput,
   Modal,
+  ImageUpload,
+  ColorPalette,
 };
